requirements.txt
astroid==3.2.3
click==8.1.7
cmake==3.30.0
colorama==0.4.6
dill==0.3.8
#dlib @dlib-19.24.99-cp312-cp312-win_amd64.whl
#face-recognition @face_recognition-1.3.0-py2.py3-none-any.whl
face_recognition_models==0.3.0
isort==5.13.2
mccabe==0.7.0
numpy==1.26.4
opencv-python==4.10.0.84
pillow==10.4.0
platformdirs==4.2.2
pylint==3.2.5
setuptools==70.3.0
tomlkit==0.13.0
